fix validation and bugs

+++ src/controllers/category/controller/deleteCategory.ts
@@ -4,6 +4,12 @@ import ProductCategory from '../../../model/ProductCategory';
 export const deleteProductCategory = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const adminId = req.userId;
+    if (!adminId) {
+      return res
+        .status(403)
+        .json({ message: 'Unauthorized access: Admin ID is missing' });
+    }
     const deletedCategory = await ProductCategory.findByIdAndDelete(id);
     if (!deletedCategory) {
       return res.status(404).json({ message: 'Category not found' });
